Refactor file structure to conform to namespace and gem name.
