+++ lib/unicorn/worker_killer/configuration.rb
